DayPercent de int a double con 2 decimales en el calculo

+++ src/Application/Features/TimesControl/Queries/GetTimeControlRangeByEmployeeId/GetTimeControlRangeByEmployeeIdResponse.cs
@@ -19,5 +19,5 @@ public record GetTimeControlRangeByEmployeeIdResponse
         TimeState TimeState,
         ClosedBy ClosedBy,
         int Minutes,
-        int DayPercent);
+        double DayPercent);
 }
